requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
binary==1.0.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.16.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
nose==1.3.7
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
PyJWT==1.4.2
pytest==4.1.1
pytest-cov==2.6.1
six==1.12.0
Werkzeug==0.14.1
